Rename AddToMeal.vue -> DiaryAddToMeal.vue
